[CHORE] Refactor Pets Component
